third_party/blink/web_tests/fast/js/mozilla/strict/B.1.1-expected.txt

PASS Function("'use strict'; 010") threw exception of type SyntaxError.
PASS !!Function("010") is true
PASS true === true
PASS Function("function f() { \"use strict\"; 010; }") threw exception of type SyntaxError.
PASS true === true
PASS !!Function("function f() { \"use strict\"; }; 010") is true
PASS true === true
PASS !!Function("function f() { 010; }") is true
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
